Refactored unit testing
