+++ src/funcs/__mocks__/fetch.js
@@ -1,0 +1,962 @@
+const sample = {
+    "data": [
+        {
+            "trainNumber": 8467,
+            "departureDate": "2019-10-19",
+            "operatorUICCode": 10,
+            "operatorShortCode": "vr",
+            "trainType": "HL",
+            "trainCategory": "Commuter",
+            "commuterLineID": "U",
+            "runningCurrently": false,
+            "cancelled": false,
+            "version": 263077544051,
+            "timetableType": "REGULAR",
+            "timetableAcceptanceDate": "2019-07-05T09:06:42.000Z",
+            "timeTableRows": [
+                {
+                    "stationShortCode": "HKI",
+                    "stationUICCode": 1,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "14",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:12:00.000Z"
+                },
+                {
+                    "stationShortCode": "PSL",
+                    "stationUICCode": 10,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "7",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:16:30.000Z"
+                },
+                {
+                    "stationShortCode": "PSL",
+                    "stationUICCode": 10,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "7",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:17:00.000Z"
+                },
+                {
+                    "stationShortCode": "ILA",
+                    "stationUICCode": 9,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": false,
+                    "commercialTrack": "",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:18:18.000Z"
+                },
+                {
+                    "stationShortCode": "ILA",
+                    "stationUICCode": 9,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": false,
+                    "commercialTrack": "",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:18:18.000Z"
+                },
+                {
+                    "stationShortCode": "KHK",
+                    "stationUICCode": 1028,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": false,
+                    "commercialTrack": "",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:18:28.000Z"
+                },
+                {
+                    "stationShortCode": "KHK",
+                    "stationUICCode": 1028,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": false,
+                    "commercialTrack": "",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:18:28.000Z"
+                },
+                {
+                    "stationShortCode": "HPL",
+                    "stationUICCode": 72,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:19:30.000Z"
+                },
+                {
+                    "stationShortCode": "HPL",
+                    "stationUICCode": 72,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:20:00.000Z"
+                },
+                {
+                    "stationShortCode": "VMO",
+                    "stationUICCode": 847,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": false,
+                    "commercialTrack": "",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:20:48.000Z"
+                },
+                {
+                    "stationShortCode": "VMO",
+                    "stationUICCode": 847,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": false,
+                    "commercialTrack": "",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:20:48.000Z"
+                },
+                {
+                    "stationShortCode": "PJM",
+                    "stationUICCode": 69,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": false,
+                    "commercialTrack": "",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:21:28.000Z"
+                },
+                {
+                    "stationShortCode": "PJM",
+                    "stationUICCode": 69,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": false,
+                    "commercialTrack": "",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:21:28.000Z"
+                },
+                {
+                    "stationShortCode": "MÄK",
+                    "stationUICCode": 693,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": false,
+                    "commercialTrack": "",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:22:14.000Z"
+                },
+                {
+                    "stationShortCode": "MÄK",
+                    "stationUICCode": 693,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": false,
+                    "commercialTrack": "",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:22:14.000Z"
+                },
+                {
+                    "stationShortCode": "LPV",
+                    "stationUICCode": 68,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:23:30.000Z"
+                },
+                {
+                    "stationShortCode": "LPV",
+                    "stationUICCode": 68,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:24:00.000Z"
+                },
+                {
+                    "stationShortCode": "KIL",
+                    "stationUICCode": 580,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:25:42.000Z"
+                },
+                {
+                    "stationShortCode": "KIL",
+                    "stationUICCode": 580,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:26:00.000Z"
+                },
+                {
+                    "stationShortCode": "KEA",
+                    "stationUICCode": 621,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:27:42.000Z"
+                },
+                {
+                    "stationShortCode": "KEA",
+                    "stationUICCode": 621,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:28:00.000Z"
+                },
+                {
+                    "stationShortCode": "KNI",
+                    "stationUICCode": 67,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:29:42.000Z"
+                },
+                {
+                    "stationShortCode": "KNI",
+                    "stationUICCode": 67,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:30:00.000Z"
+                },
+                {
+                    "stationShortCode": "KVH",
+                    "stationUICCode": 675,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:32:12.000Z"
+                },
+                {
+                    "stationShortCode": "KVH",
+                    "stationUICCode": 675,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:32:30.000Z"
+                },
+                {
+                    "stationShortCode": "TRL",
+                    "stationUICCode": 579,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:34:12.000Z"
+                },
+                {
+                    "stationShortCode": "TRL",
+                    "stationUICCode": 579,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:34:30.000Z"
+                },
+                {
+                    "stationShortCode": "EPO",
+                    "stationUICCode": 66,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:36:30.000Z"
+                },
+                {
+                    "stationShortCode": "EPO",
+                    "stationUICCode": 66,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:37:00.000Z"
+                },
+                {
+                    "stationShortCode": "KLH",
+                    "stationUICCode": 65,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:39:30.000Z"
+                },
+                {
+                    "stationShortCode": "KLH",
+                    "stationUICCode": 65,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:40:00.000Z"
+                },
+                {
+                    "stationShortCode": "MAS",
+                    "stationUICCode": 64,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:44:42.000Z"
+                },
+                {
+                    "stationShortCode": "MAS",
+                    "stationUICCode": 64,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:45:00.000Z"
+                },
+                {
+                    "stationShortCode": "VKH",
+                    "stationUICCode": 1300,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": false,
+                    "commercialTrack": "",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:45:59.000Z"
+                },
+                {
+                    "stationShortCode": "VKH",
+                    "stationUICCode": 1300,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": false,
+                    "commercialTrack": "",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:45:59.000Z"
+                },
+                {
+                    "stationShortCode": "JRS",
+                    "stationUICCode": 578,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:46:42.000Z"
+                },
+                {
+                    "stationShortCode": "JRS",
+                    "stationUICCode": 578,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:47:00.000Z"
+                },
+                {
+                    "stationShortCode": "HEK",
+                    "stationUICCode": 1023,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": false,
+                    "commercialTrack": "",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:49:04.000Z"
+                },
+                {
+                    "stationShortCode": "HEK",
+                    "stationUICCode": 1023,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": false,
+                    "commercialTrack": "",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:49:04.000Z"
+                },
+                {
+                    "stationShortCode": "TOL",
+                    "stationUICCode": 830,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "2",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:49:42.000Z"
+                },
+                {
+                    "stationShortCode": "TOL",
+                    "stationUICCode": 830,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "2",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:50:00.000Z"
+                },
+                {
+                    "stationShortCode": "KKN",
+                    "stationUICCode": 63,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "2",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:53:00.000Z"
+                }
+        ]
+        },
+        {
+            "trainNumber": 8469,
+            "departureDate": "2019-10-19",
+            "operatorUICCode": 10,
+            "operatorShortCode": "vr",
+            "trainType": "HL",
+            "trainCategory": "Commuter",
+            "commuterLineID": "U",
+            "runningCurrently": false,
+            "cancelled": false,
+            "version": 263077544051,
+            "timetableType": "REGULAR",
+            "timetableAcceptanceDate": "2019-07-05T09:06:42.000Z",
+            "timeTableRows": [
+                {
+                    "stationShortCode": "HKI",
+                    "stationUICCode": 1,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "14",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:42:00.000Z"
+                },
+                {
+                    "stationShortCode": "PSL",
+                    "stationUICCode": 10,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "7",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:46:30.000Z"
+                },
+                {
+                    "stationShortCode": "PSL",
+                    "stationUICCode": 10,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "7",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:47:00.000Z"
+                },
+                {
+                    "stationShortCode": "ILA",
+                    "stationUICCode": 9,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": false,
+                    "commercialTrack": "",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:48:18.000Z"
+                },
+                {
+                    "stationShortCode": "ILA",
+                    "stationUICCode": 9,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": false,
+                    "commercialTrack": "",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:48:18.000Z"
+                },
+                {
+                    "stationShortCode": "KHK",
+                    "stationUICCode": 1028,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": false,
+                    "commercialTrack": "",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:48:28.000Z"
+                },
+                {
+                    "stationShortCode": "KHK",
+                    "stationUICCode": 1028,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": false,
+                    "commercialTrack": "",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:48:28.000Z"
+                },
+                {
+                    "stationShortCode": "HPL",
+                    "stationUICCode": 72,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:49:30.000Z"
+                },
+                {
+                    "stationShortCode": "HPL",
+                    "stationUICCode": 72,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:50:00.000Z"
+                },
+                {
+                    "stationShortCode": "VMO",
+                    "stationUICCode": 847,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": false,
+                    "commercialTrack": "",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:50:48.000Z"
+                },
+                {
+                    "stationShortCode": "VMO",
+                    "stationUICCode": 847,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": false,
+                    "commercialTrack": "",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:50:48.000Z"
+                },
+                {
+                    "stationShortCode": "PJM",
+                    "stationUICCode": 69,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": false,
+                    "commercialTrack": "",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:51:28.000Z"
+                },
+                {
+                    "stationShortCode": "PJM",
+                    "stationUICCode": 69,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": false,
+                    "commercialTrack": "",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:51:28.000Z"
+                },
+                {
+                    "stationShortCode": "MÄK",
+                    "stationUICCode": 693,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": false,
+                    "commercialTrack": "",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:52:14.000Z"
+                },
+                {
+                    "stationShortCode": "MÄK",
+                    "stationUICCode": 693,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": false,
+                    "commercialTrack": "",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:52:14.000Z"
+                },
+                {
+                    "stationShortCode": "LPV",
+                    "stationUICCode": 68,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:53:30.000Z"
+                },
+                {
+                    "stationShortCode": "LPV",
+                    "stationUICCode": 68,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:54:00.000Z"
+                },
+                {
+                    "stationShortCode": "KIL",
+                    "stationUICCode": 580,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:55:42.000Z"
+                },
+                {
+                    "stationShortCode": "KIL",
+                    "stationUICCode": 580,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:56:00.000Z"
+                },
+                {
+                    "stationShortCode": "KEA",
+                    "stationUICCode": 621,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:57:42.000Z"
+                },
+                {
+                    "stationShortCode": "KEA",
+                    "stationUICCode": 621,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:58:00.000Z"
+                },
+                {
+                    "stationShortCode": "KNI",
+                    "stationUICCode": 67,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T10:59:42.000Z"
+                },
+                {
+                    "stationShortCode": "KNI",
+                    "stationUICCode": 67,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T11:00:00.000Z"
+                },
+                {
+                    "stationShortCode": "KVH",
+                    "stationUICCode": 675,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T11:02:12.000Z"
+                },
+                {
+                    "stationShortCode": "KVH",
+                    "stationUICCode": 675,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T11:02:30.000Z"
+                },
+                {
+                    "stationShortCode": "TRL",
+                    "stationUICCode": 579,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T11:04:12.000Z"
+                },
+                {
+                    "stationShortCode": "TRL",
+                    "stationUICCode": 579,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T11:04:30.000Z"
+                },
+                {
+                    "stationShortCode": "EPO",
+                    "stationUICCode": 66,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T11:06:30.000Z"
+                },
+                {
+                    "stationShortCode": "EPO",
+                    "stationUICCode": 66,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T11:07:00.000Z"
+                },
+                {
+                    "stationShortCode": "KLH",
+                    "stationUICCode": 65,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T11:09:30.000Z"
+                },
+                {
+                    "stationShortCode": "KLH",
+                    "stationUICCode": 65,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T11:10:00.000Z"
+                },
+                {
+                    "stationShortCode": "MAS",
+                    "stationUICCode": 64,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T11:14:42.000Z"
+                },
+                {
+                    "stationShortCode": "MAS",
+                    "stationUICCode": 64,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T11:15:00.000Z"
+                },
+                {
+                    "stationShortCode": "VKH",
+                    "stationUICCode": 1300,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": false,
+                    "commercialTrack": "",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T11:15:59.000Z"
+                },
+                {
+                    "stationShortCode": "VKH",
+                    "stationUICCode": 1300,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": false,
+                    "commercialTrack": "",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T11:15:59.000Z"
+                },
+                {
+                    "stationShortCode": "JRS",
+                    "stationUICCode": 578,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T11:16:42.000Z"
+                },
+                {
+                    "stationShortCode": "JRS",
+                    "stationUICCode": 578,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T11:17:00.000Z"
+                },
+                {
+                    "stationShortCode": "HEK",
+                    "stationUICCode": 1023,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": false,
+                    "commercialTrack": "",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T11:19:04.000Z"
+                },
+                {
+                    "stationShortCode": "HEK",
+                    "stationUICCode": 1023,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": false,
+                    "commercialTrack": "",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T11:19:04.000Z"
+                },
+                {
+                    "stationShortCode": "TOL",
+                    "stationUICCode": 830,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T11:19:42.000Z"
+                },
+                {
+                    "stationShortCode": "TOL",
+                    "stationUICCode": 830,
+                    "countryCode": "FI",
+                    "type": "DEPARTURE",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T11:20:00.000Z"
+                },
+                {
+                    "stationShortCode": "KKN",
+                    "stationUICCode": 63,
+                    "countryCode": "FI",
+                    "type": "ARRIVAL",
+                    "trainStopping": true,
+                    "commercialStop": true,
+                    "commercialTrack": "1",
+                    "cancelled": false,
+                    "scheduledTime": "2019-10-19T11:23:00.000Z"
+                }
+            ]
+        }
+    ],
+    "status": 200,
+    "statusText": "",
+    "headers": {
+      "content-type": "application/json;charset=UTF-8",
+      "cache-control": "max-age=180, public"
+    },
+    "config": {
+      "url": "https://rata.digitraffic.fi/api/v1/live-trains/station/HKI/kkn",
+      "method": "get",
+      "headers": {
+        "Accept": "application/json, text/plain, */*"
+      },
+      "transformRequest": [
+        null
+      ],
+      "transformResponse": [
+        null
+      ],
+      "timeout": 0,
+      "xsrfCookieName": "XSRF-TOKEN",
+      "xsrfHeaderName": "X-XSRF-TOKEN",
+      "maxContentLength": -1
+    },
+    "request": {}
+}
+
+function fetch() {
+    return Promise.resolve(sample)
+}
+  
+export default fetch;
